+++ kbe/src/server/baseapp/backuper.cpp
@@ -90,8 +90,8 @@ void Backuper::createBackupTable()
 	}
 
 	// 随机一下序列
-	const unsigned int seed = (int)std::chrono::system_clock::now().time_since_epoch().count();
-	std::shuffle(backupEntityIDs_.begin(), backupEntityIDs_.end(), std::default_random_engine(seed));
+	const long long seed = std::chrono::system_clock::now().time_since_epoch().count();
+	std::shuffle(backupEntityIDs_.begin(), backupEntityIDs_.end(), std::default_random_engine((int)seed));
 }
 
 }
